HomePage의 무한 스크롤 첫 페이지에 SSR 적용
