--- src/services/snsPost/fetchSnsPosts.ts
@@ -1,20 +1,0 @@
-import axios, { AxiosResponse } from 'axios'
-import { BACKEND_URL } from 'constants/common'
-import { SnsPostResponseAboutFiltering } from 'types/snsPost'
-
-const fetchSnsPosts = async (): Promise<
-  AxiosResponse<{ data: SnsPostResponseAboutFiltering[] }>
-> => {
-  return axios({
-    method: 'get',
-    url: `${BACKEND_URL}/api/sns-posts`,
-    params: {
-      'populate[0]': 'postImages',
-      'populate[1]': 'likeUsers',
-      'populate[2]': 'author',
-      sort: 'createdAt:desc',
-    },
-  })
-}
-
-export default fetchSnsPosts
